change tref to float

+++ driver/main.cpp
@@ -267,11 +267,11 @@ int main(int argc, char* argv[])
     }
     else if(base_arg == "getitemfp16")
     {
-        drv = new GetitemDriver<float16, float16>();
+        drv = new GetitemDriver<float16, float>();
     }
     else if(base_arg == "getitembfp16")
     {
-        drv = new GetitemDriver<bfloat16, bfloat16>();
+        drv = new GetitemDriver<bfloat16, float>();
     }
     else
     {
